moved browser detection to base
