+++ lexos/static/js/scripts_upload.js
@@ -194,56 +194,4 @@ $(function () {
   if (window.File && window.FileList && window.FileReader) {
     Init()
   }
-
-  browserDetection()
 })
-
-/**
- * Display a warning in an alert if using an unsupported browser. Supported
- * browsers include Chrome and Firefox.
- */
-function browserDetection () {
-
-  const browserWarningMsg = `You are currently using an unsupported browser.\
-  Some Lexos features may not function properly or be unavailable. Please\
-  use either Chrome or Firefox when using Lexos.`
-
-  // Opera 8.0+
-  const isOpera = (!!window.opr && !!opr.addons) || !!window.opera || navigator.userAgent.indexOf(' OPR/') >= 0;
-
-  // Firefox 1.0+
-  const isFirefox = typeof InstallTrigger !== 'undefined';
-
-  // Safari 3.0+ "[object HTMLElementConstructor]"
-  const isSafari = /constructor/i.test(window.HTMLElement) || (function (p) { return p.toString() === "[object SafariRemoteNotification]"; })(!window['safari'] || (typeof safari !== 'undefined' && safari.pushNotification));
-
-  // Internet Explorer 6-11
-  const isIE = /*@cc_on!@*/false || !!document.documentMode;
-
-  // Edge 20+
-  const isEdge = !isIE && !!window.StyleMedia;
-
-  // Chrome 1+
-  const isChrome = !!window.chrome && !!window.chrome.webstore;
-
-  // If first time on page, show error message if browser is not supported
-  if (typeof(Storage) !== "undefined") {
-    // check browser if user has not previously visited page
-    if (!sessionStorage.visited){
-      if (isOpera) {
-        alert(browserWarningMsg)
-      }
-      else if (isSafari) {
-        alert(browserWarningMsg)
-      }
-      else if (isIE){
-        alert(browserWarningMsg)
-      }
-      else if(isEdge){
-        alert(browserWarningMsg)
-      }
-      // Set visited to true after the user visits the page
-      sessionStorage.visited = true
-    }
-  }
-}
